reels-automation/scripts/requirements.txt
obs-websocket-py==1.0
pyautogui==0.9.54
pyyaml==6.0.1
requests==2.31.0
pillow==10.1.0
